Move types into files
