+++ samples/SampleApi/SampleApi/ConfigureCosmosOptions.cs
@@ -1,0 +1,15 @@
+using System.Text.Json;
+using Atc.Cosmos;
+using Microsoft.Extensions.Options;
+
+namespace SampleApi;
+
+public class ConfigureCosmosOptions : IConfigureOptions<CosmosOptions>
+{
+    public void Configure(CosmosOptions options)
+    {
+        options.UseCosmosEmulator();
+        options.DatabaseName = "SampleApi";
+        options.SerializerOptions.PropertyNamingPolicy = JsonNamingPolicy.CamelCase;
+    }
+}
