+++ samples/SampleApi/SampleApi/FooContainerInitializer.cs
@@ -1,0 +1,18 @@
+using Atc.Cosmos;
+using Microsoft.Azure.Cosmos;
+
+namespace SampleApi;
+
+public class FooContainerInitializer : ICosmosContainerInitializer
+{
+    public Task InitializeAsync(
+        Database database,
+        CancellationToken cancellationToken) =>
+        database.CreateContainerIfNotExistsAsync(
+            new ContainerProperties
+            {
+                PartitionKeyPath = "/pk",
+                Id = "foo",
+            },
+            cancellationToken: cancellationToken);
+}
